sniper-bot/requirements.txt
web3==6.16.0
websockets==12.0
python-dotenv==1.0.1
